requirements-tox.txt
# dependencies for tox environment
-e .

fastapi==0.110.0
hyperopt==0.2.7
imbalanced_learn==0.12.0
joblib==1.3.2
matplotlib==3.8.3
mlflow==2.11.1
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.4
pytest==8.1.1
PyYAML==6.0.1
scikit_learn==1.4.1.post1
setuptools==69.1.1
streamlit==1.31.1
uvicorn==0.28.0
